Modularize JavaScript Output and Update Binder
Modularize the JavaScript output and update the bindings to match. This
has two important advantages:

* Modularizing the JS output namespaces the classes to prevent name
collisions when including multiple libraries.
* The JS module has a callback that is called when the runtime is ready
that fixes a problem where your code tries to use the Emscripten classes
before they are initialized.

This also includes an update to the sample to use the JS init callback.